+++ ios/HybridVideoConfig.swift
@@ -1,4 +1,5 @@
 import Foundation
+import NitroModules
 
 class HybridVideoConfig: HybridVideoConfigSpec {
   func configureAutoplay(config: AutoplayConfig) throws {
@@ -16,5 +17,28 @@ class HybridVideoConfig: HybridVideoConfigSpec {
     DispatchQueue.main.async {
       AudioSessionManager.isManagementEnabled = enabled
     }
+  }
+
+  func configureCache(config: CacheConfig) throws {
+    if let maxSizeBytes = config.maxSizeBytes, maxSizeBytes > 0 {
+      VideoCache.shared.maxSizeBytes = Int64(maxSizeBytes)
+      VideoCache.shared.enforceLimit()
+    }
+  }
+
+  func clearCache() throws -> Promise<Void> {
+    let promise = Promise<Void>()
+    VideoCache.shared.clear {
+      promise.resolve(withResult: ())
+    }
+    return promise
+  }
+
+  func getCacheSizeBytes() throws -> Promise<Double> {
+    let promise = Promise<Double>()
+    VideoCache.shared.totalSize { size in
+      promise.resolve(withResult: Double(size))
+    }
+    return promise
   }
 }